Object-Oriented interface + Example
Added an object-oriented interface for the library with a small example.

Added AdsHandle
Added AdsVariable
Added AdsNotification
Added AdsDevice

AdsLibOOI is meant to be a second library layer on top of AdsLib. The
idea is to have just an object oriented abstraction layer on top of
the existing libary, with the benefit, that we can use TcAdsDll on
TwinCAT systems for the lower layer!